The great type flip over.
All types are now defined at the beginning of lit so that we can
define the renderer type at the beginning. This allows to have
functions needing a renderer be part of the module they belong to and
avoid module name duplication. E.g. instead of Lit.Renderer.Buf.map we
will have Lit.Buf.map and thus a single Buf module (modulo the private
one for renderer implementation). This should improve readability
both for lit's implementation and for clients.